\name{expit}
\alias{expit}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
expit function
}
\description{
Performs the expit transformation
}
\usage{
expit(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
a numeric matrix/vector/scalar
}
}
\value{
each value of x is return as expit of the original value
}
