\name{logit}
\alias{logit}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
logit function
}
\description{
Performs the logit transformation
}
\usage{
logit(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
a numeric matrix/vector/scalar
}
}
\value{
each value of x is return as logit of the original value
}
